requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
faiss-cpu>=1.7.4
google-generativeai>=0.7.0
pypdf>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
tiktoken>=0.7.0
rank-bm25>=0.2.2